Our First Submission Revised
Our first submission using CF algorithm revised.
Note: I trained the model on the full train dataset since there is no
need to split to train and test and evaluate the model becuase the
target in the train dataset is binary while the one we need to predict
for the competition is float.
This have scored 0.625