Forth Submission using Decision Tree model
This is the code that i ran on EC2 instance. It applies the Decision
Tree algorithm on the full_train and full_test datasets.
Note: Model was trained on full train dataset without splitting to train
and test because the target value in train is different than the one we
submit (Binary vs Float).